+++ types/chain.go
@@ -269,3 +269,14 @@ type HardForkFeature struct {
 	Rfc         string `json:"rfc"`
 	EpochNumber uint64 `json:"epoch_number,omitempty"`
 }
+
+type TransactionProof struct {
+	Proof         *Proof `json:"proof"`
+	BlockHash     Hash   `json:"block_hash"`
+	WitnessesRoot Hash   `json:"witnesses_root"`
+}
+
+type Proof struct {
+	Indices []uint `json:"indices"`
+	Lemmas  []Hash `json:"lemmas"`
+}
