--- types/raw_tx_pool.go
@@ -1,6 +1,0 @@
-package types
-
-type RawTxPool struct {
-	Pending  []Hash `json:"pending"`
-	Proposed []Hash `json:"proposed"`
-}
